cliente unificado para que no haya errores
